Moved general parallelization code
